Refactor validation and file handling logic in services

+++ PersonnelInfo/src/main/java/org/work/personnelinfo/personel/service/PersonelService.java
@@ -13,28 +13,28 @@ import org.work.personnelinfo.resourceFile.service.ResourceFileService;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class PersonelService {
-
+    private static final String PERSONEL_NOT_FOUND_WITH_ID = "Personel not found with id: ";
     private final PersonelRepository personelRepository;
     private final PersonelMapper personelMapper;
     private final ResourceFileService resourceFileService;
 
-    @Transactional(readOnly = true)
-    public PersonelDTO getPersonelById(Long id) {
-        if (id == null) {
-            throw new IllegalArgumentException("Id cannot be null");
-        }
-        return personelRepository.findById(id)
-                .map(personelMapper::modelToDTO)
-                .orElseThrow(() -> new EntityNotFoundException("Personel not found with id: " + id));
+    private enum ProcessType {
+        UPLOAD, DELETE
     }
 
     @Transactional(readOnly = true)
-    public List<PersonelDTO> getPersonelAll() {
+    public PersonelDTO getPersonelById(Long id) {
+        return personelMapper.modelToDTO(getPersonelEntityById(id));
+    }
+
+    @Transactional(readOnly = true)
+    public List<PersonelDTO> getAllPersonel() {
         List<PersonelEntity> personelEntities = personelRepository.findAll();
         return personelEntities.stream()
                 .map(personelMapper::modelToDTO)
@@ -43,60 +43,38 @@ public class PersonelService {
 
     @Transactional
     public PersonelDTO addPersonel(PersonelDTO personelDTO, MultipartFile file) throws IOException {
-        if (personelDTO == null ) {
-            throw new IllegalArgumentException("PersonelDTO cannot be null");
-        }
-
-        PersonelEntity personelEntity = personelMapper.dtoToModel(personelDTO);
-
-        personelEntity = personelRepository.save(personelEntity);
-
-        if (file != null && !file.isEmpty()) {
-            resourceFileService.uploadFile(file, personelEntity);
-        }
-
-        return personelMapper.modelToDTO(personelEntity);
+        PersonelEntity newPersonelEntity = personelMapper.dtoToModel(Objects.requireNonNull(personelDTO, "PersonelDTO cannot be null"));
+        newPersonelEntity = personelRepository.save(newPersonelEntity);
+        handleFile(ProcessType.UPLOAD, file, newPersonelEntity);
+        return personelMapper.modelToDTO(newPersonelEntity);
     }
 
     @Transactional
     public PersonelDTO updatePersonel(Long personelId, PersonelDTO personelDTO, MultipartFile file) throws IOException {
-        if (personelId == null || personelDTO == null) {
-            throw new IllegalArgumentException("PersonelId or PersonelDTO cannot be null");
-        }
-
-        PersonelEntity existingPersonelEntity = personelRepository.findById(personelId)
-                .orElseThrow(() -> new EntityNotFoundException("Personel not found with id: " + personelId));
-
-        personelMapper.updateModel(personelDTO, existingPersonelEntity);
-
-        if (file != null && !file.isEmpty()) {
-            if (existingPersonelEntity.getResourceFile() != null) {
-                resourceFileService.deleteFile(existingPersonelEntity.getResourceFile().getId());
-            }
-            resourceFileService.uploadFile(file, existingPersonelEntity);
-        }
-
+        PersonelEntity existingPersonelEntity = getPersonelEntityById(Objects.requireNonNull(personelId, "PersonelId cannot be null"));
+        personelMapper.updateModel(Objects.requireNonNull(personelDTO, "PersonelDTO cannot be null"), existingPersonelEntity);
+        handleFile(ProcessType.DELETE, file, existingPersonelEntity);
         PersonelEntity updatedPersonelEntity = personelRepository.save(existingPersonelEntity);
-
         return personelMapper.modelToDTO(updatedPersonelEntity);
     }
 
     @Transactional
-    public void deletePersonel(Long personelId) throws FileNotFoundException {
-        if (personelId == null) {
-            throw new IllegalArgumentException("PersonelId cannot be null");
-        }
-
-        PersonelEntity personelEntity = personelRepository.findById(personelId)
-                .orElseThrow(() -> new EntityNotFoundException("Personel not found with id: " + personelId));
-
-        if (personelEntity.getResourceFile() != null) {
-            resourceFileService.deleteFile(personelEntity.getResourceFile().getId());
-        }
-
+    public void deletePersonel(Long personelId) throws IOException {
+        PersonelEntity personelEntity = getPersonelEntityById(Objects.requireNonNull(personelId, "PersonelId cannot be null"));
+        handleFile(ProcessType.DELETE, null, personelEntity);
         personelRepository.delete(personelEntity);
     }
 
+    private void handleFile(ProcessType processType, MultipartFile file, PersonelEntity personelEntity) throws IOException {
+        if (file != null && !file.isEmpty()) {
+            if (processType == ProcessType.DELETE && personelEntity.getResourceFile() != null) {
+                resourceFileService.deleteFile(personelEntity.getResourceFile().getId());
+            }
+            resourceFileService.saveFile(file, personelEntity);
+        }
+    }
 
-
+    private PersonelEntity getPersonelEntityById(Long id) {
+        return personelRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(PERSONEL_NOT_FOUND_WITH_ID + id));
+    }
 }
